Router optimizations
