+++ src/js/index.js
@@ -1,4 +1,1 @@
-import Router from './api/router';
-import * as operators from './operators';
-
-export { Router, operators };
+export { Router, RouterCore } from './api/router';
